feat(parser): provide parse context { line, column, data }
Be able to access the current parse context in #onStartTag, #onEndTag
and #onError. This allows users to provide more precise error messages.

As computing the context is somewhat expensive, users must explicitly
invoke a getter to retrieve it. The context replaces the previously
existing #getStringNode getter.

This commit also cleans up error messages and adds tests for them.