feat(parser): support anonymous namespaces
This allow anonymous namespace elements to be imported without
an error being thrown.

The parser makes no correction of the prefix according to a
defined document-wide targetNamespace. It is up to the
sophisticated client library to perform that adjustment.

This partically reverts e876f1fbcaffe2c50505ab66dca3ec4f6199cfd2.